Section68[Chap508..509]-Delete Customer: extended Controller and adapted Service

+++ spring-json-rest-demo/src/main/java/com/luv2code/springdemo/service/CustomerServiceImpl.java
@@ -44,7 +44,11 @@ public class CustomerServiceImpl implements CustomerService {
 	@Override
 	@Transactional
 	public void deleteCustomer(int theId) {
-		
+		Customer foundCust = customerDAO.getCustomer(theId);
+
+		if (foundCust == null) {
+			throw new CustomerNotFoundException("Customer with ID="+theId+" not found!");
+		}
 		customerDAO.deleteCustomer(theId);
 	}
 }
